+++ tests/style/naming_enumerations/Test_1.m
@@ -1,0 +1,19 @@
+% (c) Copyright 2021 Florian Schanda
+
+classdef Test_1
+
+    enumeration
+        foo  % no
+        Bar  % ok
+    end
+
+    enumeration
+        BAZ         % ok
+        Bork_Class  % ok
+    end
+
+    enumeration
+        wrong5  % no
+    end
+
+end